requirements.txt
fastapi
uvicorn
pydantic
asyncio
jinja2
sympy
google-generativeai
Pillow
